+++ xenon-common/src/main/java/com/vmware/xenon/common/StatelessService.java
@@ -293,11 +293,34 @@ public class StatelessService implements Service {
         return s;
     }
 
+    /**
+     * Value indicating whether GET on /available returns 200 or 503
+     * The method is a convenience method since it relies on STAT_NAME_AVAILABLE to report
+     * availability.
+     */
+    public void setAvailable(boolean isAvailable) {
+        this.toggleOption(ServiceOption.INSTRUMENTATION, true);
+        this.setStat(STAT_NAME_AVAILABLE, isAvailable ? 1.0 : 0.0);
+    }
+
+    /**
+     * Value indicating whether GET on /available returns 200 or 503
+     */
+    public boolean isAvailable() {
+        if (!hasOption(Service.ServiceOption.INSTRUMENTATION)) {
+            return true;
+        }
+        ServiceStat st = this.getStat(STAT_NAME_AVAILABLE);
+        if (st != null && st.latestValue == 1.0) {
+            return true;
+        }
+        return false;
+    }
+
     private boolean allocateUtilityService() {
         synchronized (this.options) {
             if (this.utilityService == null) {
                 this.utilityService = new UtilityService().setParent(this);
-                ;
             }
         }
         return true;